Mapper class extracted
